% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportGEFAnnotations.R
\name{ImportGEFAnnotations}
\alias{ImportGEFAnnotations}
\title{Importing Histological Annotations}
\usage{
ImportGEFAnnotations(SectionName, InputAnnotationFile)
}
\description{
ImportHistologicalAnnotations()
}
