% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportHistologicalAnnotations.R
\name{ImportHistologicalAnnotations}
\alias{ImportHistologicalAnnotations}
\title{Importing Histological Annotations}
\usage{
ImportHistologicalAnnotations(SectionName, InputAnnotationFile)
}
\description{
ImportHistologicalAnnotations()
}
