% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportHistologicalOriginalSTSelections.R
\name{ImportHistologicalOriginalSTSelections}
\alias{ImportHistologicalOriginalSTSelections}
\title{ImportHistologicalOriginalSTSelections}
\usage{
ImportHistologicalOriginalSTSelections(SectionName, InputAnnotationFile)
}
\description{
ImportHistologicalOriginalSTSelections()
}
