% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportOriginalSTCountData.R
\name{ImportOriginalSTCountData}
\alias{ImportOriginalSTCountData}
\title{Importing Histological Annotations}
\usage{
ImportOriginalSTCountData(SectionName, InputCountFile)
}
\description{
ImportHistologicalAnnotations()
}
