% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergingCountAndAnnotationData.R
\name{MergingCountAndAnnotationData}
\alias{MergingCountAndAnnotationData}
\title{Importing Histological Annotations}
\usage{
MergingCountAndAnnotationData(SectionName, InputAnnotationFile, InputCountFile)
}
\description{
ImportHistologicalAnnotations()
}
