% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output_PGA_Visualization_MatrixGreyNA.R
\name{Output_PGA_Visualization_MatrixGreyNA}
\alias{Output_PGA_Visualization_MatrixGreyNA}
\title{Output_PGA_Visualization_MatrixGreyNA}
\usage{
Output_PGA_Visualization_MatrixGreyNA(SectionName, InputCNVs, BarcodesFile)
}
\description{
Output_PGA_Visualization_MatrixGreyNA()
}
