% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractSectionWise.R
\name{ExtractSectionWise}
\alias{ExtractSectionWise}
\title{ExtractSectionWise}
\usage{
ExtractSectionWise(
  SectionName,
  CNV_Genes_Organscale_Input,
  AllBarcodes,
  Threshold
)
}
\description{
ExtractSectionWise()
}
