% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalAnnotations.R
\name{FinalAnnotations}
\alias{FinalAnnotations}
\title{Importing Histological Annotations}
\usage{
FinalAnnotations(InputOriginalAnnotationFile, InputCounts)
}
\description{
ImportHistologicalAnnotations()
}
