% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OriginalST_MergingCountAndAnnotationData.R
\name{OriginalST_MergingCountAndAnnotationData}
\alias{OriginalST_MergingCountAndAnnotationData}
\title{Importing Histological Annotations}
\usage{
OriginalST_MergingCountAndAnnotationData(InputAnnotationFile, InputCountFile)
}
\description{
ImportHistologicalAnnotations()
}
