% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_PGA_Visualization_Matrix().R
\name{Plot_PGA_Visualization_Matrix}
\alias{Plot_PGA_Visualization_Matrix}
\title{Plotting the spatial distribution of genes with an inferred copy number alteration from an underlying matrix}
\usage{
Plot_PGA_Visualization_Matrix(SectionName, InputMatrix, MaxValInput)
}
\description{
Plot_PGA_Visualization_Matrix()
}
