% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectingSubTreeData.R
\name{SelectingSubTreeData}
\alias{SelectingSubTreeData}
\title{Selecting Subtree Data for Node Selection}
\usage{
SelectingSubTreeData(SubtreeObject, NodeOfInterest)
}
\description{
SelectingSubTreeData()
}
